eventbookingsystem/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
django-filter==21.1
Django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.1
Markdown==3.3.6
pip==21.1.2
psycopg2==2.9.3
pytz==2021.3
setuptools==57.0.0
sqlparse==0.4.2
wheel==0.36.2
zipp==3.7.0
